--- a/src/full/Agda/TypeChecking/Irrelevance.hs-boot
+++ b/src/full/Agda/TypeChecking/Irrelevance.hs-boot
@@ -8,7 +8,6 @@ import Agda.TypeChecking.Monad.Debug (MonadDebug)
 import {-# SOURCE #-} Agda.TypeChecking.Pretty (PrettyTCM)
 import Agda.TypeChecking.Monad.Pure (PureTCM)
 
-workOnTypes :: (MonadTCEnv m, HasOptions m, MonadDebug m) => m a -> m a
 isPropM
   :: (LensSort a, PrettyTCM a, PureTCM m, MonadBlock m)
   => a -> m Bool