--- a/src/full/Agda/TypeChecking/Monad.hs
+++ b/src/full/Agda/TypeChecking/Monad.hs
@@ -8,6 +8,7 @@ module Agda.TypeChecking.Monad
     , module Agda.TypeChecking.Monad.Env
     , module Agda.TypeChecking.Monad.Imports
     , module Agda.TypeChecking.Monad.MetaVars
+    , module Agda.TypeChecking.Monad.Modality
     , module Agda.TypeChecking.Monad.Mutual
     , module Agda.TypeChecking.Monad.Open
     , module Agda.TypeChecking.Monad.Options
@@ -29,6 +30,7 @@ import Agda.TypeChecking.Monad.Debug
 import Agda.TypeChecking.Monad.Env
 import Agda.TypeChecking.Monad.Imports
 import Agda.TypeChecking.Monad.MetaVars
+import Agda.TypeChecking.Monad.Modality
 import Agda.TypeChecking.Monad.Mutual
 import Agda.TypeChecking.Monad.Options
 import Agda.TypeChecking.Monad.Open