@andreasabel 2023-04-30
Thanks for the PR, @SquidDev !

I think the module `Agda.TypeChecking.Irrelevance` has outgrown its original scope (namely irrelevance), now it covers other modalities as well.
I looked at your separation of this module; it seems that not all functions that do not use reductions have been moved out.
I would suggest the split criterion to be: 
1. does not use reduction vs. (everything until including `wakeIrrelevantVars`?)
2. does use reduction (everything starting `UsableAtModality`?).

Could the module 1. maybe be named `Agda.TypeChecking.Monad.Modality` (rather than ... `Rules.WorkOnTypes`)?
The second could maybe keep the current name, unless there is some better suggestion.