@SquidDev 2023-05-06
Oh, that readthedocs failure is very odd. Will try to reproduce locally, not sure what's caused that.